server/pyproject.toml
[project]
name = "ecommerce-server"
version = "0.1.0"
description = "FastAPI E-commerce Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "motor>=3.4.0",
    "python-jose[cryptography]>=3.3.0",
    "match>=0.3.2", # dependency needed for python 3.12 sometimes, but let's just use argon2
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.1.1",
    "stripe>=9.0.0",
    "cloudinary>=1.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
    "black>=24.4.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.black]
line-length = 88
target-version = ['py312']
